refactor(advisory-details): remove dead CSAF legacy components
Delete csaf-vulnerability-card.tsx and csaf-remediations.tsx.
Both were fully replaced by csaf-vulnerabilities-tab/ components and had
zero imports.

Merge orphaned linkifyDetails tests into the existing csaf-utils test
file alongside collectProducts tests.
