--- client/src/app/pages/advisory-details/components/csaf-vulnerability-card.tsx
@@ -1,324 +1,0 @@
-import React from "react";
-import { Link } from "react-router-dom";
-
-import dayjs from "dayjs";
-
-import {
-  Button,
-  Card,
-  CardBody,
-  CardHeader,
-  CardTitle,
-  Content,
-  DescriptionList,
-  DescriptionListDescription,
-  DescriptionListGroup,
-  DescriptionListTerm,
-  Divider,
-  ExpandableSection,
-  Flex,
-  FlexItem,
-  Label,
-  Stack,
-  StackItem,
-  Title,
-} from "@patternfly/react-core";
-import AngleRightIcon from "@patternfly/react-icons/dist/esm/icons/angle-right-icon";
-import ExternalLinkAltIcon from "@patternfly/react-icons/dist/esm/icons/external-link-alt-icon";
-
-import { Paths } from "@app/Routes";
-import { SeverityShieldAndText } from "@app/components/SeverityShieldAndText";
-import type {
-  JSONSchemaForCommonVulnerabilityScoringSystemVersion30,
-  JSONSchemaForCommonVulnerabilityScoringSystemVersion31,
-  Vulnerability,
-} from "@app/specs/csaf/csaf-v2.0-schema";
-
-import { CsafCvssDetails } from "./csaf-cvss-details";
-import { CsafRemediations } from "./csaf-remediations";
-
-const expandableSectionStyle = {
-  "--pf-v6-c-expandable-section--m-expanded__toggle-icon--Rotate": "90deg",
-} as React.CSSProperties;
-
-type CvssV3 =
-  | JSONSchemaForCommonVulnerabilityScoringSystemVersion30
-  | JSONSchemaForCommonVulnerabilityScoringSystemVersion31;
-
-interface CsafVulnerabilityCardProps {
-  vulnerability: Vulnerability;
-  productNameMap: Map<string, string>;
-}
-
-const INITIAL_PRODUCTS_SHOWN = 5;
-
-const AffectedProducts: React.FC<{
-  productIds: string[];
-  productNameMap: Map<string, string>;
-}> = ({ productIds, productNameMap }) => {
-  const [expanded, setExpanded] = React.useState(false);
-  const visible = expanded
-    ? productIds
-    : productIds.slice(0, INITIAL_PRODUCTS_SHOWN);
-  const remaining = productIds.length - INITIAL_PRODUCTS_SHOWN;
-
-  return (
-    <Flex gap={{ default: "gapXs" }} flexWrap={{ default: "wrap" }}>
-      {visible.map((id) => (
-        <FlexItem key={id}>
-          <Label variant="outline" color="orange" isCompact>
-            {productNameMap.get(id) ?? id}
-          </Label>
-        </FlexItem>
-      ))}
-      {!expanded && remaining > 0 && (
-        <FlexItem>
-          <Button variant="link" size="sm" onClick={() => setExpanded(true)}>
-            +{remaining} more
-          </Button>
-        </FlexItem>
-      )}
-      {expanded && productIds.length > INITIAL_PRODUCTS_SHOWN && (
-        <FlexItem>
-          <Button variant="link" size="sm" onClick={() => setExpanded(false)}>
-            Show less
-          </Button>
-        </FlexItem>
-      )}
-    </Flex>
-  );
-};
-
-export const CsafVulnerabilityCard: React.FC<CsafVulnerabilityCardProps> = ({
-  vulnerability,
-  productNameMap,
-}) => {
-  const cvss = vulnerability.scores?.[0]?.cvss_v3 as CvssV3 | undefined;
-  const affectedProducts = vulnerability.product_status?.known_affected || [];
-  const [cvssExpanded, setCvssExpanded] = React.useState(true);
-
-  const cveLink = vulnerability.cve?.startsWith("CVE-")
-    ? Paths.vulnerabilityDetails.replace(":vulnerabilityId", vulnerability.cve)
-    : null;
-
-  return (
-    <Card data-testid="vulnerability-card">
-      <CardHeader>
-        <CardTitle style={{ fontWeight: 400 }}>
-          <Flex
-            alignItems={{ default: "alignItemsCenter" }}
-            gap={{ default: "gapMd" }}
-          >
-            <FlexItem>
-              {cveLink ? (
-                <Link
-                  to={cveLink}
-                  style={{
-                    fontSize: "var(--pf-v6-global--FontSize--lg)",
-                    textDecoration: "underline",
-                  }}
-                >
-                  {vulnerability.cve}
-                </Link>
-              ) : (
-                <span style={{ fontSize: "var(--pf-v6-global--FontSize--lg)" }}>
-                  {vulnerability.cve}
-                </span>
-              )}
-            </FlexItem>
-            {cvss && (
-              <FlexItem>
-                <SeverityShieldAndText
-                  value={
-                    cvss.baseSeverity.toLowerCase() as
-                      | "critical"
-                      | "high"
-                      | "medium"
-                      | "low"
-                      | "none"
-                  }
-                  score={cvss.baseScore}
-                  showLabel
-                  showScore
-                />
-              </FlexItem>
-            )}
-          </Flex>
-        </CardTitle>
-      </CardHeader>
-
-      <CardBody>
-        <Stack hasGutter>
-          {vulnerability.title && (
-            <StackItem>
-              <Content component="p">{vulnerability.title}</Content>
-            </StackItem>
-          )}
-
-          <StackItem>
-            <DescriptionList
-              isCompact
-              isHorizontal
-              columnModifier={{ default: "2Col" }}
-            >
-              {vulnerability.cwe && (
-                <DescriptionListGroup>
-                  <DescriptionListTerm>CWE</DescriptionListTerm>
-                  <DescriptionListDescription>
-                    {vulnerability.cwe.id}: {vulnerability.cwe.name}
-                  </DescriptionListDescription>
-                </DescriptionListGroup>
-              )}
-              {cvss && (
-                <DescriptionListGroup>
-                  <DescriptionListTerm>CVSS score</DescriptionListTerm>
-                  <DescriptionListDescription>
-                    {cvss.baseScore} / 10
-                  </DescriptionListDescription>
-                </DescriptionListGroup>
-              )}
-              {vulnerability.discovery_date && (
-                <DescriptionListGroup>
-                  <DescriptionListTerm>Discovered</DescriptionListTerm>
-                  <DescriptionListDescription>
-                    {dayjs(vulnerability.discovery_date).format("MMM D, YYYY")}
-                  </DescriptionListDescription>
-                </DescriptionListGroup>
-              )}
-              {vulnerability.release_date && (
-                <DescriptionListGroup>
-                  <DescriptionListTerm>Released</DescriptionListTerm>
-                  <DescriptionListDescription>
-                    {dayjs(vulnerability.release_date).format("MMM D, YYYY")}
-                  </DescriptionListDescription>
-                </DescriptionListGroup>
-              )}
-              {vulnerability.threats &&
-                vulnerability.threats.length > 0 &&
-                vulnerability.threats.map((threat, i) => (
-                  <DescriptionListGroup key={`${threat.category}-${i}`}>
-                    <DescriptionListTerm>
-                      {threat.category
-                        .replace(/_/g, " ")
-                        .replace(/^\w/, (c) => c.toUpperCase())}
-                    </DescriptionListTerm>
-                    <DescriptionListDescription>
-                      {threat.details}
-                    </DescriptionListDescription>
-                  </DescriptionListGroup>
-                ))}
-            </DescriptionList>
-          </StackItem>
-
-          {cvss && (
-            <>
-              <StackItem>
-                <Divider />
-              </StackItem>
-              <StackItem>
-                <ExpandableSection
-                  toggleText="CVSS v3 details"
-                  toggleIcon={<AngleRightIcon />}
-                  style={expandableSectionStyle}
-                  isExpanded={cvssExpanded}
-                  onToggle={(_e, expanded) => setCvssExpanded(expanded)}
-                >
-                  <div
-                    style={{
-                      marginTop: "var(--pf-v6-global--spacer--sm)",
-                    }}
-                  >
-                    <CsafCvssDetails cvss={cvss} />
-                  </div>
-                </ExpandableSection>
-              </StackItem>
-            </>
-          )}
-
-          {affectedProducts.length > 0 && (
-            <>
-              <StackItem>
-                <Divider />
-              </StackItem>
-              <StackItem>
-                <ExpandableSection
-                  toggleText={`Affected products (${affectedProducts.length})`}
-                  toggleIcon={<AngleRightIcon />}
-                  style={expandableSectionStyle}
-                >
-                  <div
-                    style={{
-                      marginTop: "var(--pf-v6-global--spacer--sm)",
-                    }}
-                  >
-                    <AffectedProducts
-                      productIds={affectedProducts}
-                      productNameMap={productNameMap}
-                    />
-                  </div>
-                </ExpandableSection>
-              </StackItem>
-            </>
-          )}
-
-          {vulnerability.remediations &&
-            vulnerability.remediations.length > 0 && (
-              <>
-                <StackItem>
-                  <Divider />
-                </StackItem>
-                <StackItem>
-                  <CsafRemediations
-                    remediations={vulnerability.remediations}
-                    productNameMap={productNameMap}
-                  />
-                </StackItem>
-              </>
-            )}
-
-          {vulnerability.references && vulnerability.references.length > 0 && (
-            <>
-              <StackItem>
-                <Divider />
-              </StackItem>
-              <StackItem>
-                <Title headingLevel="h4" size="md">
-                  References
-                </Title>
-                <Flex
-                  direction={{ default: "column" }}
-                  gap={{ default: "gapXs" }}
-                  style={{
-                    marginTop: "var(--pf-v6-global--spacer--sm)",
-                  }}
-                >
-                  {vulnerability.references.map((ref) => (
-                    <FlexItem key={ref.url}>
-                      <a
-                        href={ref.url}
-                        target="_blank"
-                        rel="noopener noreferrer"
-                        style={{
-                          display: "inline-flex",
-                          alignItems: "center",
-                          gap: 4,
-                        }}
-                      >
-                        {ref.summary}
-                        <ExternalLinkAltIcon
-                          style={{
-                            fontSize: "var(--pf-v6-global--FontSize--xs)",
-                          }}
-                        />
-                      </a>
-                    </FlexItem>
-                  ))}
-                </Flex>
-              </StackItem>
-            </>
-          )}
-        </Stack>
-      </CardBody>
-    </Card>
-  );
-};
